+++ components/backends/trtllm/src/dynamo/trtllm/main.py
@@ -13,16 +13,17 @@ from tensorrt_llm.llmapi.tokenizer import tokenizer_factory
 
 from dynamo.llm import (
     ModelType,
-    get_tensorrtllm_engine,
-    get_tensorrtllm_publisher,
     register_llm,
 )
 from dynamo.runtime import DistributedRuntime, dynamo_worker
 from dynamo.runtime.logging import configure_dynamo_logging
-from dynamo.trtllm.utils.request_handlers.handlers import (
+from dynamo.trtllm.request_handlers.handlers import (
     RequestHandlerConfig,
     RequestHandlerFactory,
 )
+from dynamo.trtllm.engine import get_llm_engine
+from dynamo.trtllm.publisher import get_publisher
+
 from dynamo.trtllm.utils.trtllm_utils import (
     Config,
     cmd_line_args,
@@ -129,7 +130,7 @@ async def init(runtime: DistributedRuntime, config: Config):
     # We already detokenize inside HandlerBase. No need to also do it in TRTLLM.
     default_sampling_params.detokenize = False
 
-    async with get_tensorrtllm_engine(engine_args) as engine:
+    async with get_llm_engine(engine_args) as engine:
         endpoint = component.endpoint(config.endpoint)
 
         if is_first_worker(config):
@@ -159,7 +160,7 @@ async def init(runtime: DistributedRuntime, config: Config):
             kv_listener = runtime.namespace(config.namespace).component(
                 config.component
             )
-            async with get_tensorrtllm_publisher(
+            async with get_publisher(
                 component,
                 engine,
                 kv_listener,
