+++ components/backends/trtllm/src/dynamo/trtllm/request_handlers/handler_base.py
@@ -20,8 +20,8 @@ from enum import Enum
 from tensorrt_llm import SamplingParams
 from tensorrt_llm.llmapi import DisaggregatedParams as LlmDisaggregatedParams
 
-from dynamo.llm.tensorrtllm.engine import TensorRTLLMEngine
-from dynamo.llm.tensorrtllm.publisher import Publisher
+from dynamo.trtllm.engine import TensorRTLLMEngine
+from dynamo.trtllm.publisher import Publisher
 from dynamo.runtime.logging import configure_dynamo_logging
 from dynamo.trtllm.utils.disagg_utils import (
     DisaggregatedParams,
