+++ components/backends/trtllm/src/dynamo/trtllm/request_handlers/handlers.py
@@ -3,7 +3,7 @@
 
 import copy
 
-from dynamo.trtllm.utils.request_handlers.handler_base import (
+from dynamo.trtllm.request_handlers.handler_base import (
     DisaggregationMode,
     DisaggregationStrategy,
     HandlerBase,
