+++ components/backends/trtllm/src/dynamo/trtllm/utils/trtllm_utils.py
@@ -4,7 +4,7 @@
 import argparse
 from typing import Optional
 
-from dynamo.trtllm.utils.request_handlers.handler_base import (
+from dynamo.trtllm.request_handlers.handler_base import (
     DisaggregationMode,
     DisaggregationStrategy,
 )
